annoying stuff
gave up on making a neat way to search so did a messy one(but it works).
also goes to the first page on load or selection.  moved thumbnail css
from bootstrap to in page to make thumbnail backing be the same height
so its a grid and not a mess. maybe other stuff i forgot... it is late.
=(^o^)=